Break apart tests.
